+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Si vous souhaitez voir le détail, cliquez sur les numéros qui vous renverrons 
 - meilleure priorisation de l'écoute du clavier, notamment pour le QuickAccess (évite les conflits avec les autres raccourcis clavier de Weda)
 - [#611](https://github.com/Refhi/Weda-Helper/issues/611) - le message d'erreur est désormais bien plus explicite en cas d'option erronée dans les options de Weda-Helper
 - [#600](https://github.com/Refhi/Weda-Helper/issues/600) - le mémo raccourcis clavier s'affiche de nouveau correctement au bout d'une seconde d'appuis sur Alt, et ce, pour toutes les plate-formes.
+- [#609](https://github.com/Refhi/Weda-Helper/issues/609) - La carte vitale n'est lue automatiquement que sur l'onglet actuellement utilisé.
 
 # [2.16.3.2] - fix
 ## Fix :